Check if queryString is not nil before passing it to regex matching.
The “numberOfMatchesInString:options:range:” method crashes if nil is passed as the string parameter.

This could happen if the requst is a PUT/POST and the [NSString initWithData:encoding:] fails for some reason (according to NSString docs “ Returns nil if the initialization fails for some reason (for example if data does not represent valid data for encoding”).