Major changes
There seems to be a bug where randomly launchEnvironment aren’t properly injected to application, therefore the bonjour architecture that signals server to be up and running has been removed. Instead we are now polling from the SBTUITunneledApplication -> server and wait for it to start responding (using the ping command). Easy peasy.

There was a data race on startupBlockCompleted that was resolved with an NSLock