Fix NotificationCenter requestAuthorizationWithOptions
At the moment `swz_requestAuthorizationWithOptions` uses
`_autorizationStatus` to determine the `granted` boolean value for its
completion block. As it is calculated by casting the string value to
integer and then to `UNAuthorizationStatus`, the empty string gives us
`UNAuthorizationStatusNotDetermined`. According to documentation, the
default value is supposed to be `UNAuthorizationStatusAuthorized`.

By setting default value in `SBTUITestTunnelServer.notificationCenterStubbedAuthorizationStatus`
we get the expected authorisation status.