Fix threading issue in IPC connection
This change addresses the possibility of
`DTXIPCConnection._slaveDidConnectWithName` being called from non-main
thread by dispatching the work to main thread when it is being called
from non main thread. This is necessary so that `_otherConnection`
gets assigned on main thread.

`SBTUITestTunnelClient.serverDidConnect` is updated to wait for
`ipcConnection` to become valid. This is needed because
`DTXIPCConnection._slaveDidConnectWithName` and `SBTUITestTunnelClient.serverDidConnect`
can be called from different threads upon which we dispatch both to
main. But `SBTUITestTunnelClient.serverDidConnect` work can start
first at which point it can lead to assertion of `ipcConnection` not
being valid as `ipcConnection._otherConnection` has not yet been
assigned.